Edit:  moved selected game dtos to domain layer
