+++ src/modules/selectedGame/domain/dtos/IRequestActivityDTO.ts
