+++ src/modules/selectedGame/infra/repositories/RequestsRepository.ts
@@ -1,5 +1,5 @@
-import IRequestAchievementDTO from 'modules/selectedGame/dtos/IRequestAchievementDTO';
-import IRequestActivityDTO from 'modules/selectedGame/dtos/IRequestActivityDTO';
+import IRequestAchievementDTO from 'modules/selectedGame/domain/dtos/IRequestAchievementDTO';
+import IRequestActivityDTO from 'modules/selectedGame/domain/dtos/IRequestActivityDTO';
 import IRequestsRepository from 'modules/selectedGame/repositories/IRequestsRepository';
 import makeHttpProvider from 'shared/infra/providers/factories/makeHttpProvider';
