+++ src/modules/selectedGame/repositories/IRequestsRepository.ts
@@ -1,5 +1,5 @@
-import IRequestAchievementDTO from 'modules/selectedGame/dtos/IRequestAchievementDTO';
-import IRequestActivityDTO from 'modules/selectedGame/dtos/IRequestActivityDTO';
+import IRequestAchievementDTO from 'modules/selectedGame/domain/dtos/IRequestAchievementDTO';
+import IRequestActivityDTO from 'modules/selectedGame/domain/dtos/IRequestActivityDTO';
 
 export default interface IRequestsRepository {
   achievement(payload: IRequestAchievementDTO): Promise<void>;
